+++ modules/registration/routes/enrollmentRouter.js
@@ -2,7 +2,6 @@ import { Router } from "express";
 import {
   addEnrollmentDetail,
   deleteEnrollmentDetail,
-  getSemesterByStudentId,
   getOrCreateEnrollmentHead,
   getPaymentStatus,
 } from "../controllers/enrollmentController.js";
@@ -12,7 +11,6 @@ enrollmentRouter.get("/", (req, res) => {
   return res.send("Enroll");
 });
 enrollmentRouter.get("/payment/:headId", getPaymentStatus);
-enrollmentRouter.get("/semesters/:studentId", getSemesterByStudentId);
 enrollmentRouter.post("/head", getOrCreateEnrollmentHead);
 enrollmentRouter.post("/", addEnrollmentDetail);
 enrollmentRouter.delete("/:selectedEnrollmentId", deleteEnrollmentDetail);
