+++ modules/registration/routes/regisRouter.js
@@ -4,6 +4,7 @@ import { courseRouter } from "./courseRouter.js";
 import { enrollmentRouter } from "./enrollmentRouter.js";
 import { transcriptRouter } from "./transcriptRouter.js";
 import { gpaxRouter } from "./gpaxRouter.js";
+import { semesterRouter } from "./semesterRouter.js";
 const regisRouter = Router();
 
 regisRouter.get("/", (req, res) => {
@@ -14,5 +15,6 @@ regisRouter.use("/course", courseRouter);
 regisRouter.use("/enroll", enrollmentRouter);
 regisRouter.use("/transcript", transcriptRouter);
 regisRouter.use("/gpax", gpaxRouter);
+regisRouter.use("/semesters", semesterRouter);
 
 export { regisRouter };
